Improve return of delete calls
The partial delete (HTTP 207) didn't not raise an exception but returned
a list of items with the delete status and message, while not being to
delete any item raised an exception (because HTTP 400 is returned). For
coherence, no exception is now raised when no item can be deleted and
the result is the same as partial delete.